Fix -Wsign-compare diagnostics
Cast the signed results of e.g., PyString_Size to size_t.

In the event that the result is an error value somehow the negative
value will almost certainly exceed the requested size and return
with an error.

+++ fuseparts/_fusemodule.c
@@ -687,14 +687,14 @@ read_func(const char *path, char *buf, size_t s, off_t off)
 	if(PyObject_CheckBuffer(v)) {
 		PyObject_GetBuffer(v, &buffer, PyBUF_SIMPLE);
 
-		if(buffer.len <= s) {
+		if((size_t)buffer.len <= s) {
 			memcpy(buf, buffer.buf, buffer.len);
 			ret = buffer.len;
 		}
 
 		PyBuffer_Release(&buffer);
 	} else if(PyBytes_Check(v)) {
-		if(PyBytes_Size(v) > s)
+		if((size_t)PyBytes_Size(v) > s)
 			goto OUT_DECREF;
 		memcpy(buf, PyBytes_AsString(v), PyBytes_Size(v));
 		ret = PyBytes_Size(v);
@@ -948,7 +948,7 @@ getxattr_func(const char *path, const char *name, char *value, size_t size)
         /* If the size of the value buffer is too small to hold the result,  errno
          * is set to ERANGE.
          */
-		if (PyString_Size(v) > size) {
+		if ((size_t)PyString_Size(v) > size) {
             ret = -ERANGE;
 			goto OUT_DECREF;
         }
@@ -984,7 +984,7 @@ listxattr_func(const char *path, char *list, size_t size)
 	}
 
 	for (;;) {
-		int ilen;
+		size_t ilen;
 
 	        w = PyIter_Next(iter);
 		if (!w) {
